Moving mock folder from src to src/test
