+++ src/test/Provisioning/Service/CognitoIdentityService.js
@@ -4,8 +4,8 @@ import chai from 'chai';
 import {CognitoIdentityService} from '../../../lib.compiled/Provisioning/Service/CognitoIdentityService';
 import Core from 'deep-core';
 import {ObjectStorage} from 'deep-core/lib.compiled/Generic/ObjectStorage';
-import {PropertyInstanceMock} from '../../../mock/Property/PropertyInstanceMock.js';
-import {ProvisioningInstanceMock} from '../../../mock/Provisioning/ProvisioningInstanceMock';
+import {PropertyInstanceMock} from '../../mock/Property/PropertyInstanceMock';
+import {ProvisioningInstanceMock} from '../../mock/Provisioning/ProvisioningInstanceMock';
 
 
 suite('Provisioning/Service/CognitoIdentityService', function() {
